Eliminate global layout state.
Global state in this instance was preventing the ability for Sprity
to process multiple PNG streams within the same invocation of node.
The 'default' layout became a catchall for every PNG running through
the stream.

My solution was to attach `layout` to the object, which is unique
to the stream.